+++ test/integration/shop/api/payment_status_confirm_test.rb
@@ -35,7 +35,8 @@ class Shop::Api::PaymentStatusConfirmTest < ActionDispatch::IntegrationTest
           "Success" => true,
           "ErrorCode" => "0",
           "Status" => "CONFIRMED",
-          "PaymentId" => payment_id.to_s
+          "PaymentId" => payment_id.to_s,
+          "Amount" => 20_000
         }
       end
     end