+++ test/services/payments/tbank_adapter_amount_test.rb
@@ -34,7 +34,7 @@ class Payments::TbankAdapterAmountTest < ActiveSupport::TestCase
     assert_not Payments::TbankAdapter.notification_amount_matches?(@payment, { "Amount" => 100 })
   end
 
-  test "notification_amount_matches skips when Amount absent" do
-    assert Payments::TbankAdapter.notification_amount_matches?(@payment, {})
+  test "notification_amount_matches rejects when Amount absent" do
+    assert_not Payments::TbankAdapter.notification_amount_matches?(@payment, {})
   end
 end